Expose gpgme_data_set_file_name
When using `gpg -dv <filename>` to decrypt a file, it displays "original
file name" that shows the name of the original file before encrypting.
But when encrypting from GPGME::Data stream, the file name was lost. And
gpg displays "-&20" or similar string as the original file name.

This commit adds the ability to set the `file_name` property of
GPGME::Data object, this property will be displayed as "original file
name" when decrypting with `gpg -dv` command.