Disable running tests when configuring gpgme (#125)
This disables more tests when building gpgme. This adds a significant
amount of installation time and it also currently prevents this gem from
being installed on MacOS.

On MacOS, it fails with a message like the following:

Gem::Ext::BuildError: ERROR: Failed to build gem native extension.

With this change, installing the gpgme gem succeeds.