update-indentation-type-inconsistency
This PR should only contain white space changes.
I noticed some inconsistency in the use of tabs verses spaces for the indentation. Some editors display them slightly differently causing the indentation to appear oddly.

I replaced each tab with 8 spaces because that seemed to line up similar to the way github was displaying the code.

Feel free to ignore if this is incorrect or too petty of a request