Improve checks for pkg-config and fall back to gpgme-config
https://github.com/ueno/ruby-gpgme/pull/166 changed from `pkg-config`
to `gpgme-config`, but if `PKG_CONFIG_PATH` is not set properly, the C
extension library will quietly build without the right linker
dependencies.

To fix this, we now check for the existence of `gpgme.pc` in
`PKG_CONFIG_PATH` and attempt to fall back to `gpgme-config` if it
does not exist.

Closes https://github.com/ueno/ruby-gpgme/issues/167